Final Commit
Finally all the stuff is working. There are still a few functionalities that need to be implemented. However, for the current work in 2D is sufficient. I would like to use the following script to produce better model for Lukas and to upgrade it for a simple 3D model setup