A few bug correction.
I found a few bug, but  I fixed. Now, I will start doing a small script to read the data base and create publishing quality figures. In addition, I will choose whenever to separate plotting routine from the post processing routine.
I introduced some bash script to handle the submission of the python script. Here I tried to create a parallel array, but I forgot that many functions within Python are parallel as well, so it is difficult to create an array of test for me.